% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_function.R
\name{analyze_function}
\alias{analyze_function}
\title{Analyze function}
\usage{
analyze_function(fun, par, parname, datasets, ...)
}
\arguments{
\item{fun}{function to analyze and find arguments for}

\item{par}{list, where item "varname" is the variable name and "value"
is another list in which to look for arguments}

\item{parname}{name of function in parameter list (NULL if
top-level function)}

\item{datasets}{array of lists; for each array item (which is a
list), item "varname" is the variable name and "value" is a
data frame in which to look for data sources of function arguments}

\item{...}{any additional arguments in which to look for
function arguments}
}
\value{
a nested list representing function representing
function dependencies.
}
\description{
Analyze function
}
\references{
Murphy L (2023). likelihood: Methods for Maximum Likelihood Estimation_. R package version 1.9, <https://CRAN.R-project.org/package=likelihood>.
}
\author{
Lora Murphy, Cary Institute of Ecosystem Studies
}
