% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anneal.R
\name{anneal}
\alias{anneal}
\title{anneal
Performs simulated annealing to find maximum likelihood
estimates for a set of parameters.}
\usage{
anneal(
  model,
  par,
  var,
  source_data,
  par_lo = NULL,
  par_hi = NULL,
  pdf,
  dep_var,
  initial_temp = 3,
  temp_red = 0.95,
  ns = 20,
  nt = 100,
  max_iter = 50000,
  min_change = 0,
  min_drops = 100,
  hessian = TRUE,
  delta = 100,
  slimit = 2,
  c = 2,
  note = "",
  show_display = TRUE,
  ...
)
}
\arguments{
\item{model}{model function to parameterize. Arguments to this function will be provided from par, var, and source_data.}

\item{par}{list of parameters for which we are using simulated annealing to find maximum likelihood estimates.  Each par element name matches an argument in a function (either model or pdf). All values listed in par must be numeric vectors. Vectors of length greater than one have each of their elements treated separately as individual parameters to estimate.}

\item{var}{list of other variables and data needed by the model and pdf functions, any type as needed. These will be kept constant}

\item{source_data}{data frame with dependent variable and associated independent variables.}

\item{par_lo}{list of lower bounds for each parameter. Each element must match an element in par in both name and size.Omitted values are assumed to be negative infinity.}

\item{par_hi}{list of upper bounds for each parameter.  Each element must match an element in par in both name and size.Omitted values are assumed to be positive infinity.}

\item{pdf}{probability density function. Make sure to use a function such as dnorm that can calculate log probability.}

\item{dep_var}{dependent variable label in source_data.}

\item{initial_temp}{temperature at which to start annealing.}

\item{temp_red}{interval by which to reduce temperature.}

\item{ns}{interval between changes in range}

\item{nt}{interval between drops in temperature}

\item{max_iter}{maximum iterations, where one iteration is one varying of each parameter to parameterize.  The annealing will run for this many iterations, unless alternate conditions have been specified (see min_change and min_drops).}

\item{min_change}{an alternate way to specify quitting conditions. This is the minimum amount of change in likelihood in min_drop number of temperature drops.  If the change is less, execution stops.}

\item{min_drops}{the companion to min_change for alternate quitting conditions. This is the number of temperature drops over which the likelihood must have changed more than min_change for execution to continue.}

\item{hessian}{if TRUE, include the standard errors and the Hessian matrix in the output.If FALSE, do not.}

\item{delta}{when calculating support limits, the number of pieces into which to divide each parameter.  This is the size of the "step" the function takes in trying to find the support limits.}

\item{slimit}{when calculating support limits, the number of likelihood units less than the optimum likelihood for which the support intervals will be calculated. 2 units is semi-standard. 1.92 units corresponds roughly to a 95% confidence interval.}

\item{c}{range reduction parameter}

\item{note}{any note to self the user wants to make.  This will be included in the output.}

\item{show_display}{show display}

\item{...}{additional arguments to model, pdf, etc. This is not a recommended way to pass additional arguments but I have chosen to support it.}
}
\value{
A list object with information on the annealing run.
}
\description{
anneal
Performs simulated annealing to find maximum likelihood
estimates for a set of parameters.
}
\references{
Murphy L (2023)._likelihood: Methods for Maximum Likelihood Estimation_. R package version 1.9, <https://CRAN.R-project.org/package=likelihood>.
}
\author{
Lora Murphy, Cary Institute of Ecosystem Studies
}
