% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anneal_custom.R
\name{anneal_custom}
\alias{anneal_custom}
\title{anneal_custom}
\usage{
anneal_custom(input_pdf, ...)
}
\arguments{
\item{input_pdf}{The name of the model to be passed to the likelihood::anneal function}

\item{...}{All arguments to be passed to the likelihood::anneal function}
}
\value{
Resulting annealing model
}
\description{
anneal_custom
}
