% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cebe_efm.R
\docType{data}
\name{cebe_efm}
\alias{cebe_efm}
\title{cebe_efm}
\format{
Data frame with three columns
\describe{
\item{species}{Species code (two letters Genus + two letters species)}
\item{psi}{Leaf water potential in MPa}
\item{kl}{Leaf hydraulic conductance measured via evaporative flux. Units : mmol m^-2 s^-1 }
}
}
\source{
Extracted from Scoffoni et al. 2011: \url{https://academic.oup.com/plphys/article/156/2/832/6108859}
}
\usage{
cebe_efm
}
\description{
\emph{Cercocarpus betuloides} evaporative flux measurements
}
\details{
Dataset contains paired measurements of leaf hydraulic conductance and leaf water potential.
}
\keyword{datasets}
