% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{compare_sp_boot}
\alias{compare_sp_boot}
\title{Pairwise Bootstrap Comparisons of Species Predictions}
\usage{
compare_sp_boot(bootstraps, conf_level = 0.95)
}
\arguments{
\item{bootstraps}{A list containing bootstrap predictions for different species based on best fitting model.}

\item{conf_level}{Confidence level for the confidence intervals (default is 0.95).}
}
\value{
A data frame with pairwise comparisons between species, including the proportion of times one species has a higher or lower prediction than another, confidence intervals for the differences, and p-values from a binomial test.
}
\description{
Pairwise Bootstrap Comparisons of Species Predictions
}
