% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineparams.R
\name{define_parsE}
\alias{define_parsE}
\title{Define parameter values}
\usage{
define_parsE(input_df)
}
\arguments{
\item{input_df}{input d with kl and psi values}
}
\value{
Parameters used in setting predicting the best fit parameters
}
\description{
Define parameter values
}
