% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlinear.R
\name{do_the_thing_linear}
\alias{do_the_thing_linear}
\title{Fits linear parameters for species' data.}
\usage{
do_the_thing_linear(input_df, model_type, plot = F)
}
\arguments{
\item{input_df}{input folder with kl and psi values}

\item{model_type}{select appropriate model type here linear}

\item{plot}{True or false for plotting model parameters}
}
\value{
Returns best fitting model parameters for each species for nonlinear fits
}
\description{
Fits linear parameters for species' data.
}
