% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnonlinear.R
\name{do_the_thing_nonlinear}
\alias{do_the_thing_nonlinear}
\title{Compute the best fit likelihood parameters for each species for each model type}
\usage{
do_the_thing_nonlinear(
  input_df,
  model_type,
  pars1,
  par_lo1,
  par_hi1,
  model_type_char = character(),
  plot = F
)
}
\arguments{
\item{input_df}{input folder with kl and psi values}

\item{model_type}{select appropriate model type here i.e., Sigmoidal, Exponentials and Logistic. See R/fitfunctions.R for functional types}

\item{pars1}{parameters set based on values in input df. See R/defineparams.R for parameter definitions.}

\item{par_lo1}{parameter values set as lower bounds for estimates. See R/defineparams.R for parameter definitions.}

\item{par_hi1}{parameter values set as upper bounds for estimates.}

\item{model_type_char}{select appropriate model type here}

\item{plot}{True or false for plotting model parameters}
}
\value{
Returns best fitting model parameters for each species for nonlinear fits
}
\description{
Compute the best fit likelihood parameters for each species for each model type
}
\details{
See r/fitfunctions.R for functional types and see R/defineparams.R for parameter definitions.
}
