% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfunctions.R
\name{Exponential}
\alias{Exponential}
\title{Define functions and parameter values}
\usage{
Exponential(A, B, psi)
}
\arguments{
\item{A}{model parameter A}

\item{B}{model parameter B}

\item{psi}{x value for the model}

\item{Xo}{Xo parameter in the model}
}
\value{
functions to be used in estimating fits for parameters
}
\description{
Define functions and parameter values
}
