% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlinear.R
\name{fit_linear}
\alias{fit_linear}
\title{fit_linear}
\usage{
fit_linear(input_df, model_type = hydrafit::Linear, plot = F, ...)
}
\arguments{
\item{input_df}{Input data frame that contains paired conductance (e.g., "kl") and leaf water potential observations}

\item{model_type}{Select appropriate model function. Here the default is linear}

\item{plot}{True or false for plotting model parameters}

\item{...}{Additional plotting parameters if plot==T (e.g., xlab, ylab)}

\item{silent}{Suppress warnings from fit}
}
\value{
Returns best fitting model parameters for each species for nonlinear fits
}
\description{
Fits linear parameters for species' data.
}
