% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitfunctions.R
\name{Exponential}
\alias{Exponential}
\alias{Logistic}
\alias{Sigmoidal}
\alias{Exponential2}
\alias{Linear}
\title{Define functions and parameter values}
\usage{
Exponential(A, B, psi)

Logistic(A, B, Xo, psi)

Sigmoidal(A, B, Xo, psi)

Exponential2(A, B, C, psi)

Linear(A, B, psi)
}
\arguments{
\item{A}{model parameter A}

\item{B}{model parameter B}

\item{psi}{x value for the model}

\item{Xo}{Xo parameter in the model}
}
\value{
Returns models for estimating parameter values
}
\description{
Define functions and parameter values
}
\details{
Used as templates for testing the fit of exponential,
   sigmoidal, logistic, exponential variant, and linear relationships.

\bold{Model types:}

\itemize{
\item Exponential
\item Exponential2
\item Logistic
\item Sigmoidal
\item Linear
}
}
