% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{get_vcov_from_anneal}
\alias{get_vcov_from_anneal}
\title{Get var-cov element for annealing}
\usage{
get_vcov_from_anneal(fit, mu_names)
}
\arguments{
\item{fit}{Model fit with variance covariance matrix}

\item{mu_names}{Character vector of variable names.}
}
\value{
Variance-covariance matrix
}
\description{
Helper to get variance covariance matrix from anneal
}
\details{
pulls vcov, aligns it to variable names for the fit to be resampled (A,B,C (or Xo)),
then ensure that the vcov meets the expectations of the multivariate resample
}
