% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeli.R
\name{likeli}
\alias{likeli}
\title{Likeli}
\usage{
likeli(model, par, var, source_data, pdf, ...)
}
\arguments{
\item{model}{model function for which to calculate likelihood.
Arguments to this function will be provided from par
and source_data.}

\item{par}{list of parameters.  If this were simulated annealing,
these would be the varying parameters.  Doesn't make too
much sense in this context but this makes it match anneal.
Each par element name matches an argument in a function
(either model or pdf). All values listed in par must be
numeric vectors.}

\item{var}{list of other variables and data needed by the model
and pdf functions, any type as needed.}

\item{source_data}{data frame with dependent variable and
associated independent variables}

\item{pdf}{probability density function.  To calculate
negative log likelihood, use a function such as
dnorm that can calculate log probability.}

\item{...}{additional arguments to be passed to analyze_function}
}
\value{
sum of the log likelihood
}
\description{
Likeli
}
