% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likeli_4_fdHess.R
\name{likeli_4_fdHess}
\alias{likeli_4_fdHess}
\title{likeli_4_fdHess}
\usage{
likeli_4_fdHess(par_2_analyze, model, par, var, source_data, pdf)
}
\arguments{
\item{par_2_analyze}{numeric vector of parameters to be analyzed}

\item{model}{Model object}

\item{par}{parameters}

\item{var}{Variables to be analyzed}

\item{source_data}{Data source}

\item{pdf}{propability density function}
}
\value{
This will transfer the values in par_2_analyze to par for transfer to likeli.
}
\description{
likeli_4_fdHess
}
