% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{print.boot_list}
\alias{print.boot_list}
\title{Print method for lists of bootstrap results}
\usage{
\method{print}{boot_list}(x, ...)
}
\arguments{
\item{x}{Object of class 'boot_list'}

\item{...}{not used}
}
\value{
Bootstrap results
}
\description{
Print method for lists of bootstrap results
}
