% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiPx.R
\name{psiPx}
\alias{psiPx}
\title{psiPX: Estimate plrc at a given maximum conductance}
\usage{
psiPx(fx_type = character())
}
\arguments{
\item{fx_type}{Function type. See ?Exponential for function options.}
}
\value{
Returns a function which takes model parameters,"px", and "max_cond_at" as parameters.
}
\description{
Creates a function that estimates the percent loss in conductance assuming maximum conductance at a chosen leaf water potential.
}
\details{
The functions within the functional are anonymous themselves, hence the "\(params)" structure.
}
\keyword{internal}
