% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{resamplePX}
\alias{resamplePX}
\title{Resample PX}
\usage{
resamplePX(fit, px = 0.5, seed, sims = 1000, psi_max = numeric())
}
\arguments{
\item{px}{Choose what percent loss in K/gs you are solving for PX (e.g., P50, P80 etc)}

\item{seed}{Seed for reproducibility.}

\item{sims}{Number of simulated values to generate}

\item{psi_max}{estimate px based on what value of psi}

\item{df}{input data frame containing best fit models, parameter estimates and their SDs}

\item{model_type}{Best fitting model}
}
\value{
Returns a list of simulations predicted PX values the length of the number `sims`.
}
\description{
Resample PX
}
\details{
Resampling is done by sampling from a normal distribution based on the mean and standard deviation of the parameter
}
