% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plots.R
\name{save_plots}
\alias{save_plots}
\title{save_plots}
\usage{
save_plots(fit.list, file = "output.pdf", force = T)
}
\arguments{
\item{fit.list}{A list of model fits}

\item{file}{The file string to save the plots to}

\item{force}{Logical to force the input to be a list of model fits}
}
\value{
A file containing the plots
}
\description{
Saves the plots from a list of model fits to a file
}
