% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scof2012}
\alias{scof2012}
\title{scof2012}
\format{
Data frame with three columns
\describe{
\item{species}{Species code (two letters Genus + two letters species)}
\item{psi}{Leaf water potential in MPa}
\item{kl}{Leaf hydraulic conductance measured via evaporative flux. Units : mmol m^-2 s^-1 }
}
}
\source{
Extracted from Scoffoni et al. 2011: \url{https://academic.oup.com/plphys/article/156/2/832/6108859}
}
\usage{
scof2012
}
\description{
Evaporative flux measurements for five species
}
\details{
Dataset contains paired measurements of leaf hydraulic conductance and leaf water potential.

The data contains vulnerability curve data for Cercocarpus betuloides, Hedera canariensis, Quercus agrifolia, Lantana camara, and Heteromeles arbutifolia
}
\keyword{datasets}
