% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(x, na.rm = T)
}
\arguments{
\item{x}{Input value(s)}

\item{na.rm}{Remove NA values. Default is TRUE.}
}
\value{
The standard error of the input value(s).
}
\description{
Standard Error
}
\details{
na.rm=T by default because the predicted parameters estimate negative values and log(negative number)=NA
}
