% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{summary.boot_list}
\alias{summary.boot_list}
\title{Summary Method for lists of bootstrap results}
\usage{
\method{summary}{boot_list}(object, ...)
}
\arguments{
\item{object}{Object of class 'boot_list'}

\item{...}{not used}
}
\value{
A summary of the bootstrap results, including species names, and pairwise comparisons.
}
\description{
Summarizes the bootstrap results for a list of bootstrapped models. This summary only applies to intergroup (e.g., among species) comparisons.
}
