% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_limits.R
\name{support_limits}
\alias{support_limits}
\title{Support limits}
\usage{
support_limits(
  model,
  par,
  var,
  source_data,
  pdf,
  par_lo = NULL,
  par_hi = NULL,
  delta = 100,
  slimit = 2
)
}
\arguments{
\item{model}{the function which contains the model statement whose parameters}

\item{par}{the parameter list as it was used in likeli and/or neighanneal,
with the best values for the varying parameters.}

\item{var}{list of other variables and data needed by the model
and pdf functions, any type as needed.}

\item{source_data}{the source data as it was passed to likeli and/or
neighanneal.}

\item{pdf}{the probability distribution function as it was passed to likeli
and/or neighanneal.}

\item{par_lo}{list containing the lower limits of the varying parameters.  Can
be a ragged array but no more than 2 dimensions.  The names of the list
members must match those in par. Any values ommitted are assumed to be
negative infinity.}

\item{par_hi}{list containing the upper limits of the varying parameters.  Can
be a ragged array but no more than 2 dimensions.  The names of the list
members must match those in par. Any values ommitted are assumed to be
infinity.}

\item{delta}{the number of pieces into which to divide each parameter.  This
is the size of the "step" the function takes in trying to find the support
limits.}

\item{slimit}{the number of likelihood units less than the optimum
likelihood for which the support intervals will be calculated.  So if
the max likelihood is -98, then for each parameter we change it in steps
of delta until we reach a likelihood of -100.}
}
\value{
support limits for the varying parameters
}
\description{
Support limits
}
