% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliertest.R
\name{vul.cur_dixon}
\alias{vul.cur_dixon}
\title{Function for working hydraulic vulnerability curve data. Tests for 
outliers via the dixon Q test function.}
\usage{
vul.cur_dixon(data, breaks_by = 0.5, test_var = character(), ...)
}
\arguments{
\item{data}{Input data frame}

\item{breaks_by}{Values to group leaf water potential into bins by. Default breaks values into bins of 0.5 MPa}

\item{test_var}{Character string. Name of column with the conductance measurement where outlier may exist}

\item{...}{Additional parameters to pass to dixon.test function. See ?outliers::dixon.test for more information}
}
\value{
Data frame with the bins of data, the Dixon test alternative hypothesis, test statistic and p-value.
}
\description{
Function for working hydraulic vulnerability curve data. Tests for 
outliers via the dixon Q test function.
}
