% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{bootPX}
\alias{bootPX}
\title{Bootstrap percent loss in conductance.}
\usage{
bootPX(df, fx_type = character(), px = 0.5, sims = 1000)
}
\arguments{
\item{df}{input data frame containing best fit models, parameter estimates and their SDs}

\item{fx_type}{what model type fit the best}

\item{px}{choose what percent loss you are solving for PX e.g., P50, P80 etc}

\item{sims}{#how many samples to generate?}
}
\value{
returns a list of sims length predicted PX values
}
\description{
Bootstrap percent loss in conductance.
}
\details{
takes the best fit model created in [fx_select()] and estimates values from a distribution based on mean and standard deviation.
}
