% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{bootPX}
\alias{bootPX}
\title{Tool for estimating confidence intervals for percent loss in conductance.}
\usage{
bootPX(
  fit,
  px = 0.5,
  psi_max,
  seed = 123,
  sims = 1000,
  pairwise = F,
  margin = c("quantile", "tdist_mean")
)
}
\arguments{
\item{fit}{Fit object(s) from `fit_vuln_curve()`}

\item{px}{Percent loss in conductance. Default is 0.5.}

\item{psi_max}{Value of psi at which to estimate maximum conductance.}

\item{seed}{Value for reproducibility}

\item{sims}{Number of simulations to run. Default is 1000.}

\item{pairwise}{Perform pairwise test among species px values.}

\item{margin}{Choice of method to estimate margin of error. Options are "tdist" for t-distribution reconstruction or "quantile" for quantile based confidence intervals. Quantile is recommended.}
}
\description{
Tool for estimating confidence intervals for percent loss in conductance.
}
\details{
This function takes the best fit model chosen by `fx_select()` and estimates confidence intervals
 by re-sampling model parameters from a multivariate normal distribution (using MASS::mvnorm) based on the variance covariance matrix of the fitted parameters.

\itemize{
   \item Annealing function info: \url{https://tinyurl.com/annealingsim}
   \item Github page: \url{https://github.com/cran/likelihood/blob/master/R/anneal.R}. See lines 742-797 specifically within the code.
}
}
