% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootPX.R
\name{bootPX}
\alias{bootPX}
\title{Tool for estimating confidence intervals for percent loss in conductance.}
\usage{
bootPX(df, fx_type = character(), px = 0.5, sims = 1000, psi_max = numeric())
}
\arguments{
\item{df}{input data frame containing best fit models, parameter estimates and their SDs}

\item{fx_type}{Best fitting model}

\item{px}{Choose what percent loss in K/gs you are solving for PX (e.g., P50, P80 etc)}

\item{sims}{Number of simulated values to generate}

\item{psi_max}{estimate px based on what value of psi}
}
\value{
Returns a list of simulations predicted PX values the length of the number `sims`.
}
\description{
Tool for estimating confidence intervals for percent loss in conductance.
}
\details{
This function takes the best fit model chosen by `fx_select()` and estimates confidence intervals
 by re-sampling model parameters from a normal distribution based on mean and standard deviation. Note that
 the standard deviation used here is estimated as the square root of a hessian matrix created when the anneal
 functions tests for the best fit model.

\itemize{
   \item Annealing function info: \url{https://tinyurl.com/annealingsim}
   \item Github page: \url{https://github.com/cran/likelihood/blob/master/R/anneal.R}. See lines 742-797 specifically within the code.
}
}
