% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cebe_efm.R
\docType{data}
\name{cebe_efm}
\alias{cebe_efm}
\title{Cercocarpus betuloides evaporative flux measurements}
\format{
Data frame with two columns
\describe{
\item{psi}{Leaf water potential in MPa}
\item{kl}{Leaf hydraulic conductance measured via evaporative flux. Units : mmol m^-2 s^-1 }
}
}
\source{
Extracted from Scoffoni et al. 2011: <https://doi-org.stanford.idm.oclc.org/10.1104/pp.111.173856>
}
\usage{
cebe_efm
}
\description{
Dataset contains paired measurements of leaf hydraulic conductance and leaf water potential.
}
\keyword{datasets}
