% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineparams.R
\name{define_pars}
\alias{define_pars}
\alias{define_parsE}
\alias{define_parsL}
\alias{define_parsS}
\alias{define_parsE2}
\alias{define_parslin}
\title{Define parameter values}
\usage{
define_pars(input_df, model_type)

define_parsE(input_df)

define_parsL(input_df)

define_parsS(input_df)

define_parsE2(input_df)

define_parslin(input_df)
}
\arguments{
\item{input_df}{Input data frame that contains paired conductance (e.g., "kl") and leaf water potential observations}

\item{model_type}{Model type to define parameter}
}
\value{
Parameter values to be used in setting predicting the best fit parameters
}
\description{
Functions used here set initial, lower and upper limits to parameter estimates
}
\details{
Functions described here define parameters for the models named below:

\bold{Model types:}

\itemize{
\item Exponential
\item Exponential2
\item Logistic
\item Sigmoidal
\item Linear
}
}
\concept{internal}
