% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitvulncurve.R
\name{estimate_pxs}
\alias{estimate_pxs}
\title{Estimate water potential at given percent loss of conductivity}
\usage{
estimate_pxs(params, px_fx, px = list(0.2, 0.5, 0.8, 0.95), max_cond_at = 0.1)
}
\arguments{
\item{params}{parameters list}

\item{px_fx}{function to calculate water potential at given percent loss of conductivity}

\item{px}{percent loss of conductivity values to estimate (default is 20, 50, 80, 95%)}

\item{max_cond_at}{water potential at which maximum conductance is assumed (default is 0.1 MPa)}
}
\description{
Estimate water potential at given percent loss of conductivity
}
