% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_select.R
\name{fx_select}
\alias{fx_select}
\title{Select the best fitting function for each species based on the lowest AICc score.
Requires that you input all of the fitting parameter data frames from [fitlinear](R/fitlinear.R) and [fit_nonlinear](R/fitnonlinear.R)}
\usage{
fx_select(.df_lin, .df_log, .df_sig, .df_exp1, .df_exp2)
}
\arguments{
\item{.df_lin}{dataframe of linear fits}

\item{.df_log}{dataframe of logistic fits}

\item{.df_sig}{dataframe of sigmoidal fits}

\item{.df_exp1}{dataframe of first exponential fits}

\item{.df_exp2}{dataframe of second exponential fits}
}
\value{
returns a dataframe with the best fit models
}
\description{
Select the best fitting function for each species based on the lowest AICc score.
Requires that you input all of the fitting parameter data frames from [fitlinear](R/fitlinear.R) and [fit_nonlinear](R/fitnonlinear.R)
}
