% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_select.R
\name{fx_select}
\alias{fx_select}
\title{Select the best fitting function for each species based on the lowest AICc score.}
\usage{
fx_select(
  linear_fits,
  logistic_fits,
  sigmoidal_fits,
  exp1_fits,
  exp2_fits,
  quiet = TRUE
)
}
\arguments{
\item{linear_fits}{dataframe of linear fits}

\item{logistic_fits}{dataframe of logistic fits}

\item{sigmoidal_fits}{dataframe of sigmoidal fits}

\item{exp1_fits}{dataframe of first exponential fits}

\item{exp2_fits}{dataframe of second exponential fits}
}
\value{
Returns a dataframe with the best fit models. Note: D here is the third parameter for Exp2 or Sigmoidal models.
}
\description{
Requires that you input all of the fitting parameter data frames from `hydrafit::fit_vuln_curve()`.
}
