% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiPx.R
\name{psiPx}
\alias{psiPx}
\title{psiPX: Estimate plrc at a given maximum conductance}
\usage{
psiPx(model_type = character())
}
\arguments{
\item{model_type}{Function type. See ?Exponential for function options.}
}
\value{
Returns a function which takes model parameters,"px", and "max_cond_at" as parameters.
}
\description{
Creates a function that estimates the percent loss in conductance assuming maximum conductance at a chosen leaf water potential.
}
\details{
The functions within the functional are anonymous themselves, hence the "\(params)" structure.
}
\examples{
# create function for estimating pX with linear formula
psi.px <- psiPx(model_type = "Linear")
A <- 2
B <- 0.5

psi_at_p50_linear <- psi.px(A=A, B=B, px = 0.5, max_cond_at = 0)

psi_at_p50_linear
}
\keyword{internal}
